Changed backend to have order details removed and just put it all in orders
