+++ data/mysql1/ib_buffer_pool
@@ -1,32 +1,37 @@
-0,394
-0,393
-0,392
-0,391
-0,390
+3,2
+10,159
+10,78
+10,163
+10,271
+10,230
+10,172
+10,149
+10,237
+10,2
+10,3
+11,7
+11,19
+11,22
+11,12
+11,16
+11,9
+11,2
+11,3
+9,5
+9,2
+9,3
+8,4
+8,7
+8,18
+8,12
+8,11
+8,2
+8,3
 38,4
 38,3
 38,2
 38,1
 38,0
-0,389
-0,388
-37,4
-37,3
-37,2
-37,1
-37,0
-0,387
-0,386
-0,278
-0,385
-0,384
-0,319
-0,318
-36,4
-36,3
-36,2
-36,1
-36,0
 35,3
 35,2
 35,1
@@ -35,47 +40,34 @@
 34,2
 34,1
 34,0
-32,35
-0,317
-0,316
-0,283
-0,243
-33,3
-33,2
-0,9
-33,1
-33,0
-3,1
-3,3
-10,243
-10,238
-10,101
-10,72
-10,23
-10,213
-10,290
-10,310
-9,7
-8,5
-8,8
-8,17
-8,16
-8,9
-8,13
 20,3
 20,2
 20,1
 19,3
 19,2
 19,1
+18,1
 12,3
+18,3
 12,2
 12,1
 11,30
 11,29
 11,28
 11,27
+14,1
 11,26
+14,3
 11,25
 11,24
+13,1
+11,1
+13,3
 11,23
+10,310
+10,290
+10,243
+10,238
+10,213
+10,101
+10,72
