--- src/main/java/com/wondacabinetinc/wondacabinetinc/businesslayer/OrderDetailsService.java
@@ -1,7 +1,0 @@
-package com.wondacabinetinc.wondacabinetinc.businesslayer;
-
-import com.wondacabinetinc.wondacabinetinc.datalayer.OrderDetails;
-
-public interface OrderDetailsService {
-    public OrderDetails addOrderDetailsToOrder(OrderDetails orderDetails, Integer orderId);
-}
