--- src/main/java/com/wondacabinetinc/wondacabinetinc/datalayer/OrderDetailsRepository.java
@@ -1,8 +1,0 @@
-package com.wondacabinetinc.wondacabinetinc.datalayer;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderDetailsRepository extends JpaRepository<OrderDetails, Integer> {
-}
