+++ src/main/java/com/wondacabinetinc/wondacabinetinc/utils/exceptions/InvalidInputException.java
@@ -1,5 +1,9 @@
 package com.wondacabinetinc.wondacabinetinc.utils.exceptions;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.UNPROCESSABLE_ENTITY)
 public class InvalidInputException extends RuntimeException {
     public InvalidInputException() {}
