+++ src/main/java/com/wondacabinetinc/wondacabinetinc/utils/exceptions/NotFoundException.java
@@ -1,5 +1,9 @@
 package com.wondacabinetinc.wondacabinetinc.utils.exceptions;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
 public class NotFoundException extends RuntimeException{
     public NotFoundException() {}
