+++ src/main/resources/db/mysql/data.sql
@@ -1,3 +1,5 @@
 INSERT IGNORE INTO orders VALUES(1, 'Done', 123321, 'Design');
 INSERT IGNORE INTO orders VALUES(2, 'Done', 456647, 'Design');
 INSERT IGNORE INTO orders VALUES(3, 'Done', 789789, 'Design');
+
+INSERT IGNORE INTO order_details VALUES(1, 'Kitchen Cabinet', 'Chocolate Brown', 'Cherry', 'Circle');
